refactor(KB-226): split prompts page into modular components
- Extract 8 components: AgentDetail, AgentTable, CoverageStats, DiffModal, ModalWrapper, PromptEditModal, PromptPlayground, ViewVersionModal
- Create custom hooks: usePrompts (data loading + rollback), useResizablePanel (drag resize)
- Move types to types.ts, utilities to utils.ts
- Main page.tsx reduced from 1067 lines to 153 lines
- No functionality changes, just code organization
