+++ admin-next/src/app/(dashboard)/prompts/components/index.ts
@@ -1,0 +1,8 @@
+export { AgentDetail } from './AgentDetail';
+export { AgentTable } from './AgentTable';
+export { CoverageStats } from './CoverageStats';
+export { DiffModal } from './DiffModal';
+export { ModalWrapper } from './ModalWrapper';
+export { PromptEditModal } from './PromptEditModal';
+export { PromptPlayground } from './PromptPlayground';
+export { ViewVersionModal } from './ViewVersionModal';
