+++ admin-next/src/app/(dashboard)/prompts/hooks/index.ts
@@ -1,0 +1,2 @@
+export { usePrompts } from './usePrompts';
+export { useResizablePanel } from './useResizablePanel';
