+++ admin-next/src/app/(dashboard)/prompts/hooks/useResizablePanel.ts
@@ -1,0 +1,40 @@
+'use client';
+
+import { useCallback, useEffect, useRef, useState } from 'react';
+
+export function useResizablePanel(initialHeight = 400, minHeight = 150, maxOffset = 300) {
+  const [height, setHeight] = useState(initialHeight);
+  const isDragging = useRef(false);
+  const containerRef = useRef<HTMLDivElement>(null);
+
+  const handleMouseDown = useCallback((e: React.MouseEvent) => {
+    e.preventDefault();
+    isDragging.current = true;
+    document.body.style.cursor = 'row-resize';
+    document.body.style.userSelect = 'none';
+  }, []);
+
+  useEffect(() => {
+    const handleMouseMove = (e: MouseEvent) => {
+      if (!isDragging.current || !containerRef.current) return;
+      const containerRect = containerRef.current.getBoundingClientRect();
+      const newHeight = e.clientY - containerRect.top;
+      setHeight(Math.max(minHeight, Math.min(newHeight, window.innerHeight - maxOffset)));
+    };
+
+    const handleMouseUp = () => {
+      isDragging.current = false;
+      document.body.style.cursor = '';
+      document.body.style.userSelect = '';
+    };
+
+    document.addEventListener('mousemove', handleMouseMove);
+    document.addEventListener('mouseup', handleMouseUp);
+    return () => {
+      document.removeEventListener('mousemove', handleMouseMove);
+      document.removeEventListener('mouseup', handleMouseUp);
+    };
+  }, [minHeight, maxOffset]);
+
+  return { height, containerRef, handleMouseDown };
+}
